Renaming and minor stuff
